Added Validation checks

+++ src/main/java/com/learn/crud/service/EmployeeServiceimpl.java
@@ -4,12 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import com.learn.crud.exceptions.CustomException;
+import com.learn.crud.repo.EmployeeRepo;
+import com.learn.crud.utility.Utility;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.learn.crud.entity.Employee;
-import com.learn.crud.repo.EmployeeRepo;
+import com.learn.crud.entity.EmployeeDTO;
 
 @Service
 public class EmployeeServiceimpl implements EmployeeService {
@@ -17,12 +19,16 @@ public class EmployeeServiceimpl implements EmployeeService {
 	@Autowired
 	EmployeeRepo employeeRepo;
 	
-	public Employee save(Employee employee) throws CustomException {
-		Boolean isExists=employeeRepo.existsByUserName(employee.getUserName());
+	@Autowired
+	Utility utility;
+	
+	public Employee save(EmployeeDTO employeeDto) throws CustomException {
+		Boolean isExists=employeeRepo.existsByUserName(employeeDto.getUserName());
 
 		if(isExists){
 			throw new CustomException("User exists");
 		}
+		Employee employee = utility.convertClientDtoToClient(employeeDto);
 		Employee emp = employeeRepo.save(employee);
 		return emp;
 	}
@@ -51,7 +57,11 @@ public class EmployeeServiceimpl implements EmployeeService {
 		 employeeRepo.deleteAll();
 	}
 	
-	public void deleteById(Long id) {
+	public void deleteById(Long id) throws CustomException {
+		Boolean isExists=employeeRepo.existsById(id);
+		if (!isExists){
+			throw new CustomException("User doesn't exist to delete");
+		}
 		employeeRepo.deleteById(id);
 	}
 }
